requirements.txt
langchain-openai 
langchain 
python-dotenv
langchain_community
langchain_groq
langchain_core
langchain_groq
langchain_openai 
langchain_community
openai
streamlit